# server/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Blog Management System"
  description: |
    A scalable Blog Management System API that enables users to register, verify their email, and
    authenticate using JWT-based authentication (access and refresh tokens). The system supports
    secure password management, including password change and email-based password reset
    functionality.

    Authenticated users can create, update, publish, and delete blog posts, as well as engage with
    content through comments. The platform enforces role-based access control (RBAC), allowing
    administrators to manage system resources while regular users interact with blog content.

    The API provides advanced querying capabilities such as pagination, filtering, sorting, and 
    searching to efficiently retrieve blog data.

    The application follows production-grade backend practices including caching, structured logging,
    automated testing, and comprehensive error handling.

    For scalable and reliable deployment, the system is containerized using Docker and supports load
    balancing, database replication, and CI/CD pipelines
  contact:
    name: Md Ismile Hosen Siam
    email: siam679ihs@gmail.com
servers:
  - description: dev
    url: http://localhost:4000/api/v1
...
paths:
...
  "/auth/sign-in":
    post:
      summary: Sign in
      description: Authenticate a user and return an access token and also set a refreshToken HttpOnly cookie.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: example@gmail.com
                password:
                  type: string
                  format: password
                  example: password1234
              required: [email, password]
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Login successful
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                        example: "/auth/sign-in"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
...
  "/comments/{id}":
    patch:
      summary: Update comment
      description: >
        Update a comment body or visibility status. Only the comment owner or an admin
        should be allowed to perform this action.
      tags: [Comment]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentRequest"
      responses:
        "200":
          description: Comment updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: comment updated
                  data:
                    $ref: "#/components/schemas/Comment"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                        example: "/api/v1/comments/507f1f77bcf86cd799439011"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    delete:
      summary: Delete comment
      description: >
        Delete a comment. Only the comment owner or an admin should be allowed to
        perform this action.
      tags: [Comment]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      responses:
        "204":
          description: Comment deleted successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # ==================================================
  # User Routes
  # ==================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
...
    Comment:
      allOf:
        - type: object
          properties:
            id:
              type: string
              example: 507f1f77bcf86cd799439011
            body:
              type: string
              example: first comment
            status:
              type: string
              enum: [hidden, public]
              example: public
            article:
              type: string
              example: 507f1f77bcf86cd799439011
            author:
              type: string
              example: 507f1f77bcf86cd799439011
          required: [id, body, status, article, author]
        - $ref: "#/components/schemas/Timestamp"
...
    UpdateCommentRequest:
      type: object
      properties:
        body:
          type: string
          example: this is a comment
        status:
          type: string
          enum: [public, hidden]
          example: public
...
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [createdAt, updatedAt]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              error:
                type: string
                example: Bad request
              message:
                type: string
                example: Validation failed
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: field name
                    message:
                      type: string
                      example: invalid input
                    in:
                      type: string
                      enum: [path, query, header, cookie, body]
                      example: body
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              error:
                type: string
                example: Internal server error
              message:
                type: string
                example: We are sorry for the inconvenience. Please try again later.
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthorized
              message:
                type: string
                example: Unauthorized.
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 403
              error:
                type: string
                example: Forbidden
              message:
                type: string
                example: You do not have permission.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 404
              error:
                type: string
                example: Not found
              message:
                type: string
                example: Requested resource not found
...
  parameters:
...
    idPathParam:
      in: path
      name: id
      required: true
      schema:
        type: string
        example: 507f1f77bcf86cd799439011